extract dataclasses for business2030
